+++ src/strapi_kit/models/__init__.py
@@ -8,7 +8,7 @@ from .config import RetryConfig, StrapiConfig
 from .content_type import ComponentListItem, ContentTypeListItem
 from .content_type import ContentTypeInfo as CTBContentTypeInfo
 from .content_type import ContentTypeSchema as CTBContentTypeSchema
-from .enums import FilterOperator, PublicationState, SortDirection
+from .enums import DocumentStatus, FilterOperator, PublicationState, SortDirection
 from .export_format import (
     ExportData,
     ExportedEntity,
@@ -61,6 +61,7 @@ __all__ = [
     "FilterOperator",
     "SortDirection",
     "PublicationState",
+    "DocumentStatus",
     # Request models - Filters
     "FilterBuilder",
     "FilterCondition",